Updated cURL
